add course revision policy

+++ app/Http/Controllers/Staff/CourseRevisionController.php
@@ -57,6 +57,8 @@ class CourseRevisionController extends Controller
      */
     public function destroy($course, CourseRevision $revision)
     {
+        $this->authorize($revision);
+
         $revision->attachments()->delete();
         $revision->delete();
